requirements.txt
numpy
pandas
scipy
matplotlib~=3.5.1
setuptools==60.6.0
fastapi~=0.73.0
uvicorn~=0.17.1
Pillow~=9.0.0
pytz~=2021.3
nexradaws~=1.1
arm_pyart
netCDF4
